Refactor scoring models and GeminiService: remove BatchScoringRequest and BatchScoringResponse from exports, change PenaltyRule points type to float, and update GeminiService to handle penalties more effectively with improved logging and payload construction.

+++ Backend/app/models/scoring/rubric.py
@@ -15,7 +15,7 @@ class RubricCategory(BaseModel):
 class PenaltyRule(BaseModel):
     code: str
     description: str
-    points: int
+    points: float
 
 class Rubric(BaseModel):
     categories: List[RubricCategory]
